+++ app/src/main/java/fr/trackoe/decheterie/model/bean/global/Depot.java
@@ -12,7 +12,7 @@ public class Depot {
     private int decheterieId;
     private long carteActiveCarteId;
     private long comptePrepayeId;
-    private float qtyTotalUDD;
+    private double qtyTotalUDD;
     private int statut;
     private boolean isSent;
 
@@ -20,7 +20,7 @@ public class Depot {
 
     }
 
-    public Depot(long id, String nom, String dateHeure, byte[] signature, int decheterieId, long carteActiveCarteId, long comptePrepayeId, float qtyTotalUDD, int statut, boolean isSent) {
+    public Depot(long id, String nom, String dateHeure, byte[] signature, int decheterieId, long carteActiveCarteId, long comptePrepayeId, double qtyTotalUDD, int statut, boolean isSent) {
         this.id = id;
         this.nom = nom;
         this.dateHeure = dateHeure;
@@ -89,11 +89,11 @@ public class Depot {
         this.comptePrepayeId = comptePrepayeId;
     }
 
-    public float getQtyTotalUDD() {
+    public double getQtyTotalUDD() {
         return qtyTotalUDD;
     }
 
-    public void setQtyTotalUDD(float qtyTotalUDD) {
+    public void setQtyTotalUDD(double qtyTotalUDD) {
         this.qtyTotalUDD = qtyTotalUDD;
     }
