+++ app/src/main/java/fr/trackoe/decheterie/ui/fragment/ApportProFragment.java
@@ -400,7 +400,7 @@ public class ApportProFragment extends Fragment {
     /*
     * get an instance of ApportProFragment
     */
-    public static ApportProFragment newInstance(long depotId, String nomInND, boolean isUsagerMenageInND, String adresseInND, String numeroCarteInND, float apportRestantInND, String uniteApportRestantInND, float totalDecompte, int accountSettingId, long comptePrepayeId) {
+    public static ApportProFragment newInstance(long depotId, String nomInND, boolean isUsagerMenageInND, String adresseInND, String numeroCarteInND, float apportRestantInND, String uniteApportRestantInND, double totalDecompte, int accountSettingId, long comptePrepayeId) {
         ApportProFragment apportProFragment = new ApportProFragment();
         Bundle args = new Bundle();
         args.putLong    (   "depotId",                  depotId                 );
@@ -410,7 +410,7 @@ public class ApportProFragment extends Fragment {
         args.putString  (   "numeroCarteInND",          numeroCarteInND         );
         args.putFloat   (   "apportRestantInND",        apportRestantInND       );
         args.putString  (   "uniteApportRestantInND",   uniteApportRestantInND  );
-        args.putFloat   (   "totalDecompte",            totalDecompte           );
+        args.putDouble  (   "totalDecompte",            totalDecompte           );
         args.putInt     (   "accountSettingId",         accountSettingId        );
         args.putLong    (   "comptePrepayeId",          comptePrepayeId         );
 
@@ -421,7 +421,7 @@ public class ApportProFragment extends Fragment {
     /*
     * get an instance of ApportProFragment
     */
-    public static ApportProFragment newInstance(long depotId, int usagerIdFromRechercherUsagerFragment, int typeCarteIdFromRechercherUsagerFragment, boolean isComeFromRechercherUsagerFragment, String nomInND, boolean isUsagerMenageInND, String adresseInND, float apportRestantInND, String uniteApportRestantInND, float totalDecompte, int accountSettingId, long comptePrepayeId) {
+    public static ApportProFragment newInstance(long depotId, int usagerIdFromRechercherUsagerFragment, int typeCarteIdFromRechercherUsagerFragment, boolean isComeFromRechercherUsagerFragment, String nomInND, boolean isUsagerMenageInND, String adresseInND, float apportRestantInND, String uniteApportRestantInND, double totalDecompte, int accountSettingId, long comptePrepayeId) {
         ApportProFragment apportProFragment = new ApportProFragment();
         Bundle args = new Bundle();
         args.putLong    (   "depotId",                                  depotId                                 );
@@ -435,7 +435,7 @@ public class ApportProFragment extends Fragment {
         //args.putString  (   "numeroCarteInND",          numeroCarteInND         );
         args.putFloat   (   "apportRestantInND",        apportRestantInND       );
         args.putString  (   "uniteApportRestantInND",   uniteApportRestantInND  );
-        args.putFloat   (   "totalDecompte",            totalDecompte           );
+        args.putDouble  (   "totalDecompte",            totalDecompte           );
         args.putInt     (   "accountSettingId",         accountSettingId        );
         args.putLong    (   "comptePrepayeId",          comptePrepayeId         );
         apportProFragment.setArguments(args);
